Extract GuildPlayer members to extensions
