+++ FredBoat/src/main/java/fredboat/audio/player/GuildPlayer.kt
@@ -25,7 +25,6 @@
 
 package fredboat.audio.player
 
-import com.google.common.collect.Lists
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack
 import com.sedmelluq.discord.lavaplayer.track.AudioTrackEndReason
@@ -37,12 +36,10 @@ import fredboat.command.music.control.VoteSkipCommand
 import fredboat.commandmeta.MessagingException
 import fredboat.commandmeta.abs.CommandContext
 import fredboat.db.api.GuildConfigService
-import fredboat.definitions.PermissionLevel
 import fredboat.definitions.RepeatMode
-import fredboat.feature.I18n
-import fredboat.perms.Permission
-import fredboat.perms.PermsUtil
-import fredboat.sentinel.*
+import fredboat.sentinel.Guild
+import fredboat.sentinel.InternalGuild
+import fredboat.sentinel.TextChannel
 import fredboat.util.TextUtils
 import fredboat.util.extension.escapeAndDefuse
 import fredboat.util.ratelimit.Ratelimiter
@@ -50,15 +47,12 @@ import fredboat.util.rest.YoutubeAPI
 import lavalink.client.player.IPlayer
 import lavalink.client.player.LavalinkPlayer
 import lavalink.client.player.event.PlayerEventListenerAdapter
-import org.apache.commons.lang3.tuple.ImmutablePair
-import org.apache.commons.lang3.tuple.Pair
 import org.bson.types.ObjectId
 import org.json.JSONObject
 import org.slf4j.LoggerFactory
 import java.util.*
 import java.util.concurrent.ConcurrentLinkedQueue
 import java.util.function.Consumer
-import kotlin.streams.toList
 
 class GuildPlayer(
         val lavalink: SentinelLavalink,
@@ -70,7 +64,7 @@ class GuildPlayer(
         youtubeAPI: YoutubeAPI
 ) : PlayerEventListenerAdapter() {
 
-    private val audioTrackProvider: ITrackProvider = SimpleTrackProvider()
+    val audioTrackProvider: ITrackProvider = SimpleTrackProvider()
     private val audioLoader = AudioLoader(ratelimiter, audioTrackProvider, audioPlayerManager, this, youtubeAPI)
     val guildId = guild.id
     val player: LavalinkPlayer = lavalink.getLink(guild.id.toString()).player
@@ -80,44 +74,12 @@ class GuildPlayer(
     private var onErrorHook: Consumer<Throwable>? = null
     @Volatile
     private var lastLoadedTrack: AudioTrackContext? = null
-    private val historyQueue = ConcurrentLinkedQueue<AudioTrackContext>()
+    val historyQueue = ConcurrentLinkedQueue<AudioTrackContext>()
 
     companion object {
         private val log = LoggerFactory.getLogger(GuildPlayer::class.java)
         private const val MAX_HISTORY_SIZE = 20
     }
-
-    val trackCount: Int
-        get() {
-            var trackCount = audioTrackProvider.size()
-            if (player.playingTrack != null) trackCount++
-            return trackCount
-        }
-
-    //Live streams are considered to have a length of 0
-    val totalRemainingMusicTimeMillis: Long
-        get() {
-            var millis = audioTrackProvider.durationMillis
-
-            val currentTrack = if (player.playingTrack != null) internalContext else null
-            if (currentTrack != null && !currentTrack.track.info.isStream) {
-                millis += Math.max(0, currentTrack.effectiveDuration - position)
-            }
-            return millis
-        }
-
-
-    val streamsCount: Long
-        get() {
-            var streams = audioTrackProvider.streamsCount().toLong()
-            val atc = if (player.playingTrack != null) internalContext else null
-            if (atc != null && atc.track.info.isStream) streams++
-            return streams
-        }
-
-
-    val currentVoiceChannel: VoiceChannel?
-        get() = guild.selfMember.voiceChannel
 
     /**
      * @return The text channel currently used for music commands.
@@ -129,12 +91,6 @@ class GuildPlayer(
             if (!guild.selfPresent) return null
             return musicTextChannelProvider.getMusicTextChannel(guild)
         }
-
-    /**
-     * @return Users who are not bots
-     */
-    val humanUsersInCurrentVC: List<Member>
-        get() = getHumanUsersInVC(currentVoiceChannel)
 
     var repeatMode: RepeatMode
         get() = if (audioTrackProvider is AbstractTrackProvider)
@@ -168,23 +124,8 @@ class GuildPlayer(
             return enabled
         }
 
-    val isQueueEmpty: Boolean
-        get() {
-            log.trace("isQueueEmpty()")
-
-            return player.playingTrack == null && audioTrackProvider.isEmpty
-        }
-
-    val trackCountInHistory: Int
-        get() = historyQueue.size
-
-    val isHistoryQueueEmpty: Boolean
-        get() = historyQueue.isEmpty()
-
     val playingTrack: AudioTrackContext?
         get() {
-            log.trace("getPlayingTrack()")
-
             return if (player.playingTrack == null && internalContext == null) {
                 audioTrackProvider.peek()
             } else internalContext
@@ -247,65 +188,6 @@ class GuildPlayer(
         activeTextChannel?.send("Something went wrong!\n${t.message}")?.subscribe()
     }
 
-    fun joinChannel(usr: Member) {
-        val targetChannel = usr.voiceChannel
-        joinChannel(targetChannel)
-    }
-
-    fun joinChannel(targetChannel: VoiceChannel?) {
-        if (targetChannel == null) {
-            throw MessagingException(I18n.get(guild).getString("playerUserNotInChannel"))
-        }
-        if (targetChannel == currentVoiceChannel) {
-            // already connected to the channel
-            return
-        }
-
-        val guild = targetChannel.guild
-        val permissions = targetChannel.ourEffectivePermissions
-
-        if (permissions hasNot Permission.VIEW_CHANNEL) {
-            val i18n = I18n.get(guild).getString("permissionMissingBot")
-            throw MessagingException("$i18n ${Permission.VIEW_CHANNEL.uiName}")
-        }
-
-        if (permissions hasNot Permission.VOICE_CONNECT && guild.selfMember.voiceChannel != targetChannel) {
-            throw MessagingException(I18n.get(guild).getString("playerJoinConnectDenied"))
-        }
-
-        if (permissions hasNot Permission.VOICE_SPEAK) {
-            throw MessagingException(I18n.get(guild).getString("playerJoinSpeakDenied"))
-        }
-
-        if (targetChannel.userLimit > 0
-                && targetChannel.userLimit <= targetChannel.members.size
-                && permissions hasNot Permission.VOICE_MOVE_OTHERS) {
-            throw MessagingException(String.format("The channel you want me to join is full!" +
-                    " Please free up some space, or give me the permission to **%s** to bypass the limit.", //todo i18n
-                    Permission.VOICE_MOVE_OTHERS.uiName))
-        }
-
-        try {
-            lavalink.getLink(guild).connect(targetChannel)
-            log.info("Connected to voice channel $targetChannel")
-        } catch (e: Exception) {
-            log.error("Failed to join voice channel {}", targetChannel, e)
-        }
-
-    }
-
-    fun leaveVoiceChannelRequest(commandContext: CommandContext, silent: Boolean) {
-        if (!silent) {
-            val currentVc = commandContext.guild.selfMember.voiceChannel
-            if (currentVc == null) {
-                commandContext.reply(commandContext.i18n("playerNotInChannel"))
-            } else {
-                commandContext.reply(commandContext.i18nFormat("playerLeftChannel", currentVc.name))
-            }
-        }
-        lavalink.getLink(guild).disconnect()
-    }
-
     fun queue(identifier: String, context: CommandContext) {
         val ic = IdentifierContext(identifier, context.textChannel, context.member)
 
@@ -331,48 +213,9 @@ class GuildPlayer(
         play()
     }
 
-    //add a bunch of tracks to the track provider
+    /** Add a bunch of tracks to the track provider */
     fun loadAll(tracks: Collection<AudioTrackContext>) {
         audioTrackProvider.addAll(tracks)
-    }
-
-    @Suppress("LocalVariableName")
-    fun getTracksInRange(start: Int, end: Int): List<AudioTrackContext> {
-        // Make mutable
-        var start_ = start
-        var end_ = end
-
-        val result = ArrayList<AudioTrackContext>()
-
-        //adjust args for whether there is a track playing or not
-
-        if (player.playingTrack != null) {
-            if (start_ <= 0) {
-                result.add(internalContext!!)
-                end_--//shorten the requested range by 1, but still start at 0, since that's the way the trackprovider counts its tracks
-            } else {
-                //dont add the currently playing track, drop the args by one since the "first" track is currently playing
-                start_--
-                end_--
-            }
-        } else {
-            //nothing to do here, args are fine to pass on
-        }
-
-        result.addAll(audioTrackProvider.getTracksInRange(start_, end_))
-        return result
-    }
-
-    /** Similar to [getTracksInRange], but only gets the trackIds */
-    fun getTrackIdsInRange(start: Int, end: Int): List<ObjectId> = getTracksInRange(start, end).stream()
-            .map { it.trackId }
-            .toList()
-
-    fun getHumanUsersInVC(vc: VoiceChannel?): List<Member> {
-        vc ?: return emptyList()
-        return vc.members.stream()
-                .filter { !it.isBot }
-                .toList()
     }
 
     override fun toString(): String {
@@ -384,44 +227,6 @@ class GuildPlayer(
             audioTrackProvider.reshuffle()
         } else {
             throw UnsupportedOperationException("Can't reshuffle " + audioTrackProvider.javaClass)
-        }
-    }
-
-    //Success, fail message
-    private suspend fun canMemberSkipTracks(member: Member, trackIds: Collection<ObjectId>): Pair<Boolean, String> {
-        if (PermsUtil.checkPerms(PermissionLevel.DJ, member)) {
-            return ImmutablePair(true, null)
-        } else {
-            //We are not a mod
-            val userId = member.id
-
-            //if there is a currently playing track, and the track is requested to be skipped, but not owned by the
-            // requesting user, then currentTrackSkippable should be false
-            var currentTrackSkippable = true
-            val playingTrack = playingTrack
-            if (playingTrack != null
-                    && trackIds.contains(playingTrack.trackId)
-                    && playingTrack.userId != userId) {
-
-                currentTrackSkippable = false
-            }
-
-            return if (currentTrackSkippable && audioTrackProvider.isUserTrackOwner(userId, trackIds)) { //check ownership of the queued tracks
-                ImmutablePair(true, null)
-            } else {
-                ImmutablePair(false, I18n.get(guild).getString("skipDeniedTooManyTracks"))
-            }
-        }
-    }
-
-    suspend fun skipTracksForMemberPerms(context: CommandContext, trackIds: Collection<ObjectId>, successMessage: String) {
-        val pair = canMemberSkipTracks(context.member, trackIds)
-
-        if (pair.left) {
-            context.reply(successMessage)
-            skipTracks(trackIds)
-        } else {
-            context.replyWithName(pair.right)
         }
     }
 
@@ -547,18 +352,6 @@ class GuildPlayer(
         player.stopTrack()
     }
 
-    fun getTracksInHistory(start: Int, end: Int): List<AudioTrackContext> {
-        val start2 = Math.max(start, 0)
-        val end2 = Math.max(end, start)
-        val historyList = ArrayList(historyQueue)
-
-        return if (historyList.size >= end2) {
-            Lists.reverse(ArrayList(historyQueue)).subList(start2, end2)
-        } else {
-            ArrayList()
-        }
-    }
-
     override fun onTrackEnd(player: IPlayer?, track: AudioTrack?, endReason: AudioTrackEndReason?) {
         log.debug("onTrackEnd({} {} {}) called", track!!.info.title, endReason!!.name, endReason.mayStartNext)
 
@@ -599,7 +392,6 @@ class GuildPlayer(
 
     /**
      * Plays the provided track.
-     *
      *
      * Silently playing a track will not trigger the onPlayHook (which announces the track usually)
      */
