+++ FredBoat/src/main/java/fredboat/command/music/control/PauseCommand.kt
@@ -25,6 +25,7 @@
 
 package fredboat.command.music.control
 
+import fredboat.audio.player.isQueueEmpty
 import fredboat.commandmeta.abs.Command
 import fredboat.commandmeta.abs.CommandContext
 import fredboat.commandmeta.abs.ICommandRestricted
